+++ Source/Classes/Extension/UI/NSLayoutConstraint+.swift
@@ -5,6 +5,8 @@
 //  Created by Zero_D_Saber on 2025/12/2.
 //
 
+import UIKit
+
 @MainActor
 public extension ZDSWrapper where T: NSLayoutConstraint {
     static func activate(@ZDArrayBuilder<NSLayoutConstraint> _ constraints: () -> [NSLayoutConstraint]) {